fix version detection
The Alternative instance for IO is such that `a <|> b` returns the
result of `a` unless running a throws an IOError, in which case `b` is
run.

What we need instead is the MaybeT behavior, where `a <|> b` evaluates
`a` and returns the result unless it evaluated to Nothing, in which case
it evaluates `b`.