black-box tests
Take the black-box test approach as far as it can go, which is
everything prior to calling executeFile (at which point the testing
Haskell process itself gets replaced).

The limitations of this approach drove me to factor out Main.operation
and Nix.arguments, which should make it possible to write unit-tests, of
a sort, though we'll still need a sandbox to provide git repos.