pin nixpkgs
Pinning the derivation to a specific revision of nixpkgs ensures a
reproducible build. If the pinned version drifts too far from the
publicly cached channels, however, this could lead to building ghc from
source rather than downloading it as a precompiled binary. We should set
up an internal Hydra server to build and cache the project.