add minimal package
Add a version of this package that leans on whatever versions of Nix and
Git the user already has installed. The assumption that Nix is already
installed is a very safe assumption to make, given the nature of this
package. The assumption that Git is installed is also very safe,
however, we cannot safely assume the user has a recent enough version of
Git to support the sorting options we're passing to `git ls-remote`. I
should adjust that call to be backwards compatible and handle the
sorting in haskell.