fallback to `git ls-remote` without --sort
The --sort option was added in git 2.18.0 and many users are likely to
have an older version. Provide a fallback mechanism to handle this.
Another approach would be to simply not include the --sort option and
always sort the results in Haskell. This could simplify the
gitTaggedVersions implementation, but I'm hesitant to do so as I suspect
it will be less efficient when dealing with a large number of tags.

https://github.com/git/git/blob/53f9a3e157dbbc901a02ac2c73346d375e24978c/Documentation/RelNotes/2.18.0.txt#L69